requirements.txt
langchain
langchain-openai
langchain-community
flask
python-dotenv
chromadb
openai
pypdf